Enabled Angular strict mode, possible bugs still there.

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -6,31 +6,38 @@ import {
   OnDestroy,
   ViewChild,
 } from '@angular/core';
-import { UntypedFormGroup } from '@angular/forms';
+import { FormGroup } from '@angular/forms';
 import { BasketService } from '../../basket/basket.service';
 import { CheckoutRoutingService } from '../checkout-routing.service';
 import { ToastrService } from 'ngx-toastr';
 import { Basket } from '../../shared/models/basket';
 import { NavigationExtras, Router } from '@angular/router';
-
-declare var Stripe;
+import {
+  Stripe,
+  StripeCardCvcElement,
+  StripeCardExpiryElement,
+  StripeCardNumberElement,
+} from '@stripe/stripe-js/types/stripe-js';
+import { loadStripe } from '@stripe/stripe-js';
+import { firstValueFrom } from 'rxjs';
+import { Address } from 'src/app/shared/models/address';
+import { OrderToCreate } from 'src/app/shared/models/order';
 
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss'],
 })
-export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
-  @Input() checkoutForm: UntypedFormGroup;
-  @ViewChild('cardNumber', { static: true }) cardNumberElement: ElementRef;
-  @ViewChild('cardExpiry', { static: true }) cardExpiryElement: ElementRef;
-  @ViewChild('cardCvc', { static: true }) cardCvcElement: ElementRef;
-  stripe: any;
-  cardNumber: any;
-  cardExpiry: any;
-  cardCvc: any;
+export class CheckoutPaymentComponent implements OnDestroy {
+  @Input() checkoutForm?: FormGroup;
+  @ViewChild('cardNumber', { static: true }) cardNumberElement?: ElementRef;
+  @ViewChild('cardExpiry', { static: true }) cardExpiryElement?: ElementRef;
+  @ViewChild('cardCvc', { static: true }) cardCvcElement?: ElementRef;
+  stripe: Stripe | null = null;
+  cardNumber?: StripeCardNumberElement;
+  cardExpiry?: StripeCardExpiryElement;
+  cardCvc?: StripeCardCvcElement;
   cardErrors: any;
-  cardHandler = this.OnChange.bind(this);
   loading = false;
   cardNumberValid = false;
   cardExpiryValid = false;
@@ -43,58 +50,53 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     private router: Router
   ) {}
 
-  ngAfterViewInit(): void {
-    this.stripe = Stripe(
+  ngOnInit(): void {
+    loadStripe(
       'pk_test_51KAZKmJZPhkM96pO3RCCWUd76PPovg1OJMx5x4Bp8zfjfVXWAizeqG9EGAneeZKoep2Z5VEkJbMHp938TdNQF0OY00SBkMaSQn'
-    );
-    const elements = this.stripe.elements();
+    ).then((stripe) => {
+      this.stripe = stripe;
+      const elements = stripe?.elements();
+      if (elements) {
+        this.cardNumber = elements.create('cardNumber');
+        this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+        this.cardNumber.on('change', (event) => {
+          this.cardNumberValid = event.complete;
+          if (event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        });
 
-    this.cardNumber = elements.create('cardNumber');
-    this.cardNumber.mount(this.cardNumberElement.nativeElement);
-    this.cardNumber.addEventListener('change', this.cardHandler);
+        this.cardExpiry = elements.create('cardExpiry');
+        this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+        this.cardExpiry.on('change', (event) => {
+          this.cardExpiryValid = event.complete;
+          if (event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        });
 
-    this.cardExpiry = elements.create('cardExpiry');
-    this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
-    this.cardExpiry.addEventListener('change', this.cardHandler);
-
-    this.cardCvc = elements.create('cardCvc');
-    this.cardCvc.mount(this.cardCvcElement.nativeElement);
-    this.cardCvc.addEventListener('change', this.cardHandler);
+        this.cardCvc = elements.create('cardCvc');
+        this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+        this.cardCvc.on('change', (event) => {
+          this.cardCvcValid = event.complete;
+          if (event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        });
+      }
+    });
   }
 
   ngOnDestroy() {
-    this.cardNumber.destroy();
-    this.cardExpiry.destroy();
-    this.cardCvc.destroy();
-  }
-
-  OnChange(event) {
-    console.log(event);
-    if (event.error) {
-      this.cardErrors = event.error.message;
-    } else {
-      this.cardErrors = null;
-    }
-    switch (event.elementType) {
-      case 'cardNumber':
-        this.cardNumberValid = event.complete;
-        break;
-      case 'cardExpiry':
-        this.cardExpiryValid = event.complete;
-        break;
-      case 'cardCvc':
-        this.cardCvcValid = event.complete;
-        break;
-    }
+    this.cardNumber?.destroy();
+    this.cardExpiry?.destroy();
+    this.cardCvc?.destroy();
   }
 
   async submitOrder() {
     this.loading = true;
     const basket = this.basketService.getCurrentBasketValue();
+    if (!basket) throw new Error('cannot get basket');
     try {
       const createdOrder = await this.createOrder(basket);
       const paymentResult = await this.confirmPaymentWithStripe(basket);
-
       if (paymentResult.paymentIntent) {
         this.basketService.deleteBasket(basket);
         const navigationExtras: NavigationExtras = { state: createdOrder };
@@ -102,36 +104,52 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
       } else {
         this.toastr.error(paymentResult.error.message);
       }
-      this.loading = false;
-    } catch (error) {
+    } catch (error: any) {
       console.log(error);
+      this.toastr.error(error.message);
+    } finally {
       this.loading = false;
     }
   }
 
-  private async confirmPaymentWithStripe(basket: Basket) {
-    return this.stripe.confirmCardPayment(basket.clientSecret, {
+  private async confirmPaymentWithStripe(basket: Basket | null) {
+    if (!basket) throw new Error('Basket is null');
+
+    const result = this.stripe?.confirmCardPayment(basket.clientSecret!, {
       payment_method: {
-        card: this.cardNumber,
+        card: this.cardNumber!,
         billing_details: {
-          name: this.checkoutForm.get('paymentForm').get('nameOnCard').value,
+          name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value,
         },
       },
     });
+
+    if (!result) throw new Error('Problem attempting payment with stripe');
+    return result;
   }
 
-  private async createOrder(basket: Basket) {
+  private async createOrder(basket: Basket | null) {
+    if (!basket) throw new Error('Basket is null');
+
     const orderToCreate = this.getOrderToCreate(basket);
-    return this.checkoutService.createOrder(orderToCreate).toPromise();
+    return firstValueFrom(this.checkoutService.createOrder(orderToCreate));
   }
 
-  private getOrderToCreate(basket: Basket) {
+  private getOrderToCreate(basket: Basket | null): OrderToCreate {
+    if (!basket) throw new Error('Basket is null');
+
+    const deliveryMethodId = this.checkoutForm
+      ?.get('deliveryForm')
+      ?.get('deliveryMethod')?.value;
+    const shipToAddress = this.checkoutForm?.get('addressForm')
+      ?.value as Address;
+
+    if (!deliveryMethodId || !shipToAddress)
+      throw new Error('Problem with basket');
     return {
       basketId: basket.id,
-      deliveryMethodId: +this.checkoutForm
-        .get('deliveryForm')
-        .get('deliveryMethod').value,
-      shipToAddress: this.checkoutForm.get('addressForm').value,
+      deliveryMethodId: deliveryMethodId,
+      shipToAddress: shipToAddress,
     };
   }
 }
